LayoutTests/java/focus-expected.txt

Test for Applet for bug 32292: Unable to focus on embedded plugins such as Flash via javascript focus().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS "appletElem"; document.activeElement === pluginElement is true
PASS "appletElemWithFallbackContents"; document.activeElement === pluginElement is true
PASS "noCodeAppletElem"; document.activeElement === pluginElement is false
PASS "noCodeAppletElemWithFallbackContents"; document.activeElement === pluginElement is false
PASS "noCodeAppletElemWithTabindex"; document.activeElement === pluginElement is true
PASS "noCodeAppletElemWithContenteditable"; document.activeElement === pluginElement is true
Test for bug 32292: "Unable to focus on embedded plugins such as Flash via javascript focus()"

This tests focusing Applets. See LayoutTests/plugins for Embed and Object elements.

         
